feat(console): move store to CDN
