+++ src/libs/cdn/src/proposals/workflows/mod.rs
@@ -4,5 +4,4 @@ pub mod delete;
 pub mod errors;
 pub mod init;
 pub mod pre_commit_assets;
-mod store;
 pub mod submit;
